refactor: move external link component
