+++ app/src/features/aim42/Aim42Content.tsx
@@ -1,6 +1,6 @@
 import { Parser, ProcessNodeDefinitions } from 'html-to-react'
 import React from 'react'
-import { ExternalLink } from '../../components/ExternalLink'
+import { ExternalLink } from '../ui'
 import './aim42.css'
 
 const isValidNode = function () {
