+++ app/src/features/layout/DefaultLayout.tsx
@@ -2,7 +2,7 @@ import { GithubOutlined } from '@ant-design/icons'
 import { Divider, Layout } from 'antd'
 import React from 'react'
 import { Link } from 'react-router-dom'
-import { ExternalLink } from '../../components/ExternalLink'
+import { ExternalLink } from '../ui'
 import { ItemEditor } from '../items'
 import { Header } from './Header'
 import { SideNav } from './SideNav'
