+++ app/src/features/ui/index.ts
@@ -3,3 +3,4 @@ export * from './AttributionCard'
 export * from './InfoBubble'
 export * from './AvatarDiv'
 export * from './Error404'
+export * from './ExternalLink'
