+++ app/src/pages/decisions-table.tsx
@@ -4,7 +4,7 @@ import { PageHeader } from '../components/PageHeader'
 import { POSSIBLE_STATUSES } from '../components/Status'
 import { ItemsTablePage } from '../features/items'
 import { AttributionCard } from '../features/ui'
-import { ExternalLink } from '../components/ExternalLink'
+import { ExternalLink } from '../features/ui'
 
 const defaultVisibleStatuses: DecisionStatus[] = ['proposed', 'accepted']
